+++ .github/workflows/weekly.yml
@@ -1,0 +1,74 @@
+name: "weekly"
+
+on:
+  schedule:
+    - cron: '0 1 * * 1'
+
+jobs:
+  LitmusAcceptance:
+    env:
+      HONEYCOMB_WRITEKEY: 7f3c63a70eecc61d635917de46bea4e6
+      HONEYCOMB_DATASET: litmus tests
+    runs-on: self-hosted
+    strategy:
+      matrix:
+        ruby_version: [2.5.x]
+        puppet_gem_version: [~> 6.0]
+        platform: [release_checks]
+        agent_family: ['puppet5', 'puppet6']
+
+    steps:
+    - uses: actions/checkout@v1
+    - name: Litmus Parallel
+      uses: puppetlabs/action-litmus_parallel@master
+      with:
+        platform: ${{ matrix.platform }}
+        agent_family: ${{ matrix.agent_family }}
+  Spec:
+    runs-on: self-hosted
+    strategy:
+      matrix:
+        check: [parallel_spec, 'syntax lint metadata_lint check:symlinks check:git_ignore check:dot_underscore check:test_file rubocop']
+        ruby_version: [2.5.x]
+        puppet_gem_version: [~> 5.0, ~> 6.0]
+        exclude:
+        - puppet_gem_version: ~> 5.0
+          check: 'syntax lint metadata_lint check:symlinks check:git_ignore check:dot_underscore check:test_file rubocop'
+        - ruby_version: 2.5.x
+          puppet_gem_version: ~> 5.0
+    steps:
+    - uses: actions/checkout@v1
+    - name: Spec Tests
+      uses: puppetlabs/action-litmus_spec@master
+      with:
+        puppet_gem_version: ${{ matrix.puppet_gem_version }}
+        check: ${{ matrix.check }}
+  
+  ComplyIntegration:
+    runs-on: self-hosted
+
+    steps:
+    - uses: actions/checkout@v1
+    - name: Build
+      run: |
+        gem install bundler
+        bundle install --path .bundle/gems --jobs 4 --retry 3
+        bundle exec rake spec_prep
+    - name: Provision and setup
+      run: |
+        export LANG=en_US.UTF-8
+        export LANGUAGE=en_US:en
+        bundle exec bolt --modulepath spec/fixtures/modules -i ./inventory.yaml plan run websphere_application_server::provision_machines pe_master=centos-7-x86_64 app_agent=oracle-5-x86_64 dmgr_agent=oracle-6-x86_64 ihs_agent=oracle-7-x86_64
+        bundle exec bolt --modulepath spec/fixtures/modules -i ./inventory.yaml plan run websphere_application_server::pe_server_setup
+        bundle exec bolt --modulepath spec/fixtures/modules -i ./inventory.yaml plan run websphere_application_server::puppet_agents_setup
+    - name: Inventory
+      run: |
+        cat ./inventory.yaml
+    - name: Run tests
+      run: |
+        env
+        bundle exec rake websphere_application_server:integration
+    - name: Tidy up
+      if: always()
+      run: |
+        bundle exec rake litmus:tear_down
