+++ metadata.json
@@ -60,5 +60,5 @@
   ],
   "pdk-version": "1.15.0",
   "template-url": "https://github.com/puppetlabs/pdk-templates#master",
-  "template-ref": "heads/master-0-gcaed9d7"
+  "template-ref": "heads/master-0-g5d52853"
 }
