+++ plans/puppet_agents_setup.pp
@@ -1,0 +1,21 @@
+plan websphere_application_server::puppet_agents_setup(
+) {
+  # get pe_server ?
+  $server = get_targets('*').filter |$n| { $n.vars['role'] == 'master' }
+
+  # get agents ?
+  $agents = get_targets('*').filter |$n| { $n.vars['role'] != 'master' }
+
+  # install agents
+  run_task('puppet_agent::install', $agents)
+
+  # set the server
+  $server_string = $server[0].name
+  run_task('puppet_conf', $agents, action => 'set', section => 'main', setting => 'server', value => $server_string)
+  
+  # run agent -t
+  run_command('puppet agent -t', $agents, '_catch_errors' => true)
+
+  # install the module
+  run_command('puppet module install puppetlabs-websphere_application_server', $agents)
+}
