--- spec/acceptance/nodesets/default.yml
@@ -1,38 +1,0 @@
-HOSTS:
-  puppet-master:
-    roles:
-      - master
-      - database
-      - dashboard
-      - agent
-    platform: el-6-x86_64
-    template: Delivery/Quality Assurance/Templates/vCloud/centos-6-x86_64
-    hypervisor: vmpooler
-  app-agent:
-    roles:
-      - agent
-      - appserver
-    platform: el-6-x86_64
-    template: Delivery/Quality Assurance/Templates/vCloud/centos-6-x86_64
-    hypervisor: vmpooler
-  dmgr-agent:
-    roles:
-      - agent
-      - dmgr
-    platform: el-6-x86_64
-    template: Delivery/Quality Assurance/Templates/vCloud/centos-6-x86_64
-    hypervisor: vmpooler
-  ihs-agent:
-    roles:
-      - agent
-      - ihs
-    platform: el-6-x86_64
-    template: Delivery/Quality Assurance/Templates/vCloud/centos-6-x86_64
-    hypervisor: vmpooler
-CONFIG:
-  nfs_server: none
-  consoleport: 443
-  datastore: instance0
-  folder: Delivery/Quality Assurance/Enterprise/Dynamic
-  resourcepool: delivery/Quality Assurance/Enterprise/Dynamic
-  pooling_api: http://vmpooler.delivery.puppetlabs.net/
