--- spec/acceptance/nodesets/ubuntu-1604-64mda.yml
@@ -1,38 +1,0 @@
-HOSTS:
-  puppet-master:
-    roles:
-      - master
-      - database
-      - dashboard
-      - agent
-    platform: ubuntu-16.04-amd64
-    template: Delivery/Quality Assurance/Templates/vCloud/ubuntu-1604-x86_64
-    hypervisor: vcloud
-  app-agent:
-    roles:
-      - agent
-      - appserver
-    platform: ubuntu-16.04-amd64
-    template: Delivery/Quality Assurance/Templates/vCloud/ubuntu-1604-x86_64
-    hypervisor: vcloud
-  dmgr-agent:
-    roles:
-      - agent
-      - dmgr
-    platform: ubuntu-16.04-amd64
-    template: Delivery/Quality Assurance/Templates/vCloud/ubuntu-1604-x86_64
-    hypervisor: vcloud
-  ihs-agent:
-    roles:
-      - agent
-      - ihs
-    platform: ubuntu-16.04-amd64
-    template: Delivery/Quality Assurance/Templates/vCloud/ubuntu-1604-x86_64
-    hypervisor: vcloud
-CONFIG:
-  nfs_server: none
-  consoleport: 443
-  datastore: instance0
-  folder: Delivery/Quality Assurance/Enterprise/Dynamic
-  resourcepool: delivery/Quality Assurance/Enterprise/Dynamic
-  pooling_api: http://vcloud.delivery.puppetlabs.net/
