--- spec/acceptance/nodesets/vagrant.yml
@@ -1,31 +1,0 @@
-HOSTS:
-  puppet-centos:
-    roles:
-      - master
-      - database
-      - dashboard
-      - agent
-    platform: el-6-x86_64
-    box : puppetlabs/centos-6.6-64-nocm
-    hypervisor : vagrant
-  app-centos:
-    roles:
-      - agent
-      - default
-    platform: el-6-x86_64
-    box : puppetlabs/centos-6.6-64-nocm
-    hypervisor : vagrant
-  dmgr-centos:
-    roles:
-      - agent
-    platform: el-6-x86_64
-    box : puppetlabs/centos-6.6-64-nocm
-    hypervisor : vagrant
-  ihs-centos:
-    roles:
-      - agent
-    platform: el-6-x86_64
-    box : puppetlabs/centos-6.6-64-nocm
-    hypervisor : vagrant
-CONFIG:
-  pe_dir: http://neptune.delivery.puppetlabs.net/3.8/ci-ready/
