+++ spec/support/examples/idempotent_resource.rb
@@ -4,7 +4,7 @@ shared_examples 'an idempotent resource' do
   end
 
   it 'runs a second time without changes' do
-    runner = BeakerAgentRunner.new
+    runner = LitmusAgentRunner.new
     second_result = runner.execute_agent_on(@agent, @manifest)
     expect(second_result.exit_code).to eq 0
   end
